+++ packages/k2-alpine/src/components/index.ts
@@ -2,4 +2,4 @@ export * from './Primitives'
 export * from './GlassView/GlassView'
 export * from './Button/Button'
 export * from './Button/CircularButton'
-export * from './Alert/useAlert'
+export * from './Alert/Alert'
